# api/app/openapi/OpenAPI_backend_En.yaml
openapi: 3.1.0
info:
  title: Dev Portal & Translation → Chroma Pipeline API
  version: 0.3.0
  description: >
    An API that provides the pipeline: Extract → Translate → (optional Writeback) → Packaging → Chroma upsert,
    plus CRUD/import operations for the portal_* tables.
    Reflects Phase J/K policies and clarifies POST idempotency / Request-ID / 429/409 responses.
...
servers:
  - url: https://api.example.com
    description: default (switch prod/dev via env overlay)
security: []  # In the prod overlay, apply security: [{ bearerAuth: [] }]
...
paths:
...
  /portal/view:
    get:
      tags: [Portal View]
      summary: List
      parameters:
        - in: query
          name: common_id
          schema: { type: integer }
        - in: query
          name: view_type
          schema: { type: string, enum: [form, kanban, list, calendar, search, graph, pivot, dashboard, tree, map] }
        - in: query
          name: model
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalViewList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal View]
      summary: Create
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalViewCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalView' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
  /portal/view/{id}:
    get:
      tags: [Portal View]
      summary: Get
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalView' } } }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Portal View]
      summary: Update (partial)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - $ref: '#/components/parameters/XRequestID'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalViewUpdate' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalView' } } }
        default: { $ref: '#/components/responses/Problem' }
  /portal/view/set_primary:
    post:
      tags: [Portal View]
      summary: Set primary view (unique within a common_id)
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SetPrimaryRequest' } } }
      responses:
        '204': { description: No Content }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
  /portal/view/bootstrap_from_common:
    post:
      tags: [Portal View]
      summary: Generate skeletons from view_common (alias endpoint)
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BootstrapViewRequest' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/BootstrapResult' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }

  ############################################
  # Portal: tab
  ############################################
  /portal/tab:
    get:
      tags: [Portal Tab]
      summary: List tabs
      parameters:
        - in: query
          name: view_id
          schema: { type: integer }
        - in: query
          name: tab_key
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalTabList' } } }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [Portal Tab]
      summary: Create
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalTabCreate' } } }
      responses:
        '201':
          description: Created
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalTab' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }
  /portal/tab/{id}:
    get:
      tags: [Portal Tab]
      summary: Get
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalTab' } } }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Portal Tab]
      summary: Update (partial)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - $ref: '#/components/parameters/XRequestID'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PortalTabUpdate' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/PortalTab' } } }
        default: { $ref: '#/components/responses/Problem' }
  /portal/tab/bulk_upsert:
    post:
      tags: [Portal Tab]
      summary: Bulk UPSERT multiple tabs
      parameters:
        - $ref: '#/components/parameters/XRequestID'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/BulkPortalTabRequest' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/BulkUpsertResult' } } }
        '409': { $ref: '#/components/responses/ConflictIdempotent' }
        default: { $ref: '#/components/responses/Problem' }

  ############################################
  # Portal: smart_button
  ############################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: Correlation ID per run. Generated by Cron/clients and attached to every request.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, minLength: 8, maxLength: 256 }
      description: >
        Idempotency key for POST/PUT. Deterministically generated from a time window and payload.
        Matching retries return 200/409 (implementation-dependent) with no side effects.
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ConflictIdempotent:
      description: Treated as replay due to idempotency
      headers:
        Idempotency-Status:
          description: If 'replayed', the request was treated as a replay
          schema: { type: string, enum: [replayed, conflict] }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        detail: { type: string }
        status: { type: integer }
      required: [title, status]
...
    CursorList:
      type: object
      properties:
        items: { type: array, items: {} }
        next_cursor: { type: string, nullable: true }

    ########################################
    # Portal model
    ########################################
...
    PortalView:
      type: object
      properties:
        id: { type: integer }
        common_id: { type: integer }
        view_type: { type: string }
        model: { type: string }
        view_name: { type: string, nullable: true }
        is_primary: { type: boolean }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalViewCreate:
      type: object
      properties:
        common_id: { type: integer }
        view_type: { type: string }
        model: { type: string }
      required: [common_id, view_type]
    PortalViewUpdate:
      type: object
      properties:
        view_name: { type: string }
        enabled: { type: boolean }
    PortalViewList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalView' } }

    ########################################
    # Portal tab
    ########################################
    PortalTab:
      type: object
      properties:
        id: { type: integer }
        view_id: { type: integer }
        tab_key: { type: string }
        tab_label_ja: { type: string, nullable: true }
        tab_label_en: { type: string, nullable: true }
        model: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PortalTabCreate:
      type: object
      properties:
        view_id: { type: integer }
        tab_key: { type: string }
        tab_label_ja: { type: string }
        tab_label_en: { type: string }
        model: { type: string }
      required: [view_id, tab_key]
    PortalTabUpdate:
      type: object
      properties:
        tab_label_ja: { type: string }
        tab_label_en: { type: string }
        model: { type: string }
    PortalTabList:
      allOf:
        - $ref: '#/components/schemas/CursorList'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/PortalTab' } }
    BulkPortalTabRequest:
      type: array
      items: { $ref: '#/components/schemas/PortalTabCreate' }

    ########################################
    # Portal smart_button
    ########################################
...
    BootstrapViewRequest:
      type: object
      properties:
        action_xmlids: { type: array, items: { type: string } }
        set_primary_from_common: { type: boolean, default: true }
      required: [action_xmlids]
    BootstrapResult:
      type: object
      properties:
        created: { type: integer }
        skipped: { type: integer }

    ########################################
    # Extract
    ########################################
...